vboot: Remove unused data at the end of the kernel buffer
The reven kernel partitions are 64MiB, but only around a quarter of that
is used, so we can chop off the rest. This is important because the
buffer is later hashed when measuring to the TPM, which is a fairly slow
operation that scales linearly with the size of the data.

BUG=b:333560930
TEST=cargo xtask check --vm-tests
TEST=cargo xtask qemu, log shows shrink from ~64M to ~16M

Change-Id: I451a88baf2f9c5b54fa45caac9adce4825cbcadd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5440301
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>