Add tools/uefi_test_tool
This is an executable package that will be used to add new VM tests. For
these new tests, the test tool will be launched first, allowing it to
make modifications to the system. Then the test tool will launch
crdyshim, and boot will proceed as normal.

The tool doesn't do anything yet, this is just the bare minimum to make
it buildable.

BUG=b:302700196
TEST=cargo xtask check

Change-Id: I4b379a901a03f690d37c98d6f5892c05ad2912a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5479907
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>