crdyboot: Add UpdateInfo::HEADER_SIZE_IN_BYTES
Use this constant for both serialization and deserialization. Using
`size_of::<UpdateInfo>` is not correct, as the struct's in
memory-representation is not the same as the serialized form. (And the
struct is repr(Rust), so it doesn't even have a fixed layout between
compilations.)

BUG=b:338423918
TEST=cargo xtask check

Change-Id: I6e7af66bd70ac2d2ca6d870cd9a48ed2d0571ec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5756148
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>