crdyboot: Rework UpdateInfo to own its data
There was a memory safety bug in the code: `UpdateInfo` stored the
device path as a reference to the data loaded with `get_variable_boxed`,
but that box was dropped before the `UpdateInfo`. This wasn't caught by
the compiler because of a bug in uefi-rs:
https://github.com/rust-osdev/uefi-rs/issues/1281

Note that this code is never actually run in crdyboot (see
`allow_firmware_updates` in main.rs, which currently always returns
false), so there should be no real-world impact from this bug.

To fix, make `UpdateInfo` store all of its data rather than referencing
external data. There is no longer any need to fully
serialize/deserialize to bytes, since the serialized bytes are stored
internally. Accessor methods are provided for the few fields that need
to be read or modified.

BUG=b:338423918
TEST=cargo xtask check

Change-Id: I1151e9736aa0dd4b8a32d79c6d11430534690f38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5757605
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>