avb: Add libavb bindgen
Compile libavb and run bindgen.

Remove stack protector as it is not available for the uefi targets.
Force -DAVB_USE_PRINTF_LOGS so that avb_printf is the only
op that needs to be implemented.
This flag appears to be the newer flag in libavb.

Tested these changes by booting crdyboot with modifications to
force it to attempt avb.

BUG=b:365802655
TEST=cargo test --package avb

Cq-Depend: chromium:5990805
Change-Id: I900a822aa5b2477cfd20ddd3155d8da04daccbbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5856083
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>