crdyboot: Conditionally copy SBAT revocations to a runtime UEFI var
If verbose logging is enabled, copy SbatLevel (which is only accessible
at boot time) to a runtime-accessible UEFI var: SbatLevelRT. This allows
SBAT revocations to be examined while the OS is running.

Note that this is not a security feature, just a dev tool for seeing
what revocations are installed.

BUG=b:361388978
TEST=Build crdyboot and copy to a DUT
TEST=Reboot, check contents of the new var:
TEST=xxd /sys/firmware/efi/efivars/SbatLevelRT-*

Change-Id: Ifc3c5e5723de8292fd86f6e583c6b8e6fd2a86bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5802606
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>