crdyboot: Remove unsafe code from load_capsules module
Now that the disk module has been updated to return the lifetime-less
variant of ScopedProtocol, we can store that value in DiskReader without
any lifetime issues. Drop the raw pointer and accompanying unsafe code.

This is part of a series of changes to switch to the new
uefi::boot/uefi::runtime modules:
https://github.com/rust-osdev/uefi-rs/blob/HEAD/docs/funcs_migration.md

BUG=b:362995016
TEST=cargo xtask check --vm-tests

Change-Id: I192e570145a220e7fbaf1403a44f179491977573
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5826167
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>