libcrdy: Alter the condition for enabling no_std
Instead of enabling std only when the `test` config is enabled, enable
it is the target_os is not `uefi`. This matches how it's configured in
crdyshim and crdyboot.

This change is needed for mockall support. Mockall requires std, but
when crdyshim/crdyboot enable the `test_util` feature in libcrdy for
tests, libcrdy itself won't be compiled with the `test` config
enabled.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: I3c7ca28fd86983fb364e6c4a2ffc73a956a8ba1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5849698
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>