crdyboot: Pass an iterator to get_update_table
This is a precursor to adding unit tests for get_update_table. The
uefi-rs API does not currently provide a way to construct the
`VariableKey` struct, which means we can't create them in unit
tests. Work around this by passing in an iterator of names and vendors
instead.

BUG=b:365161823,b:365817661
TEST=cargo xtask check

Change-Id: I0373de6bdec8113c8522005b20db0e8e481000e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5849704
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>