crdyboot: Add is_gpt_partition_entry_named function
Add the helper is_gpt_partition_entry_named function to compare
the GptPartitionType's partition_name with a CStr16.
Modify is_stateful_partition to use the function to
find the stateful partition.
Add unit tests to cover the various states the array
could be in for the comparison.
Can be used to look up partitions with names other
than STATE in the future.

BUG=b:365802655
TEST=cargo xtask check

Change-Id: I52fa43491d2ee56e73b707de54675177502df80d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5879365
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>