crdyboot: Add find_partition_by_name
Replace find_partition_handle_by_name with
find_partition_by_name and add GptPartitionEntry
to the return value.
The GptPartitionEntry is required to look up the partitions
and in future commits it will be useful to callers to avoid
having to look up the information again.

BUG=b:365802655
TEST=cargo xtask check

Change-Id: I81ec220a3eade8376544a04d1f19e204f066f6ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5900573
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>