crdyboot: Add Hd3MbrPartition to test handles
This isn't used yet, but will be used to test drives partitioned with
MBR.

This is part of a series of changes to make the tests clearer. By the
end of the series, there will be only one function to create MockUefi,
and the only parameter will be the boot drive.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: I99faa85496af27f4c6a5e5c8e709de50e49b2136
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5906375
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>