crdyboot: Unify creation of partition info and device paths
Some info is duplicated between partition info and device paths, for
example the start and end LBA. Add a function to create PartitionInfo,
and a function to create a device-path node for partitions that uses the
first function for most data.

This is part of a series of changes to make the tests clearer. By the
end of the series, there will be only one function to create MockUefi,
and the only parameter will be the boot drive.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: I2a430d49f15a2b61669961d5a20bc19a94ea858b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5906376
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>