crdyboot: Move create_mock_uefi_with_block_io higher
Minor code cleanup, this is just for clarity since the function is
closely tied to the details of DeviceKind.

This is part of a series of changes to make the tests clearer. By the
end of the series, there will be only one function to create MockUefi,
and the only parameter will be the boot drive.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: I86b58ae570e056dc4c945c9be2354c4a77a4f01a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5906377
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>