crdyboot: Add BootDrive test enum
This is used as the sole parameter of the unified mock function. For
now, only the first drive is used.

This is part of a series of changes to make the tests clearer. By the
end of the series, there will be only one function to create MockUefi,
and the only parameter will be the boot drive.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: I3b2b9d400c8f3e91289e52010cec36beafc7a2fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5908109
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>