crdyboot: Add partition_info_for_handle to unified mock
Also adjust a test with the new expected partition size.

This is part of a series of changes to make the tests clearer. By the
end of the series, there will be only one function to create MockUefi,
and the only parameter will be the boot drive.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: If0e37bf76409ae0ed45ad77cdb2a6e95732071b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5908110
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>