crdyboot: Check capsule header size
Capsule headers contain two sizes: the full size of the capsule, and the
size of the capsule header (which may be larger than the `CapsuleHeader`
type due to vendor-specific fields). In practice, the header should
always be smaller than the full capsule, but in order to ensure memory
safety, validate this at runtime.

Also update the tests to check both header and capsule size.

BUG=b:373881381
TEST=cargo xtask check

Change-Id: I52a5b3db44b034ece3bc6a8b54b77c6fce79556d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5937341
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>