libcrdy: Disallow zero-byte page allocation requests
It's not useful to have a zero-size `ScopedPageAllocation`, so return an
error if `num_bytes` is zero. This will allow other code to assume that
the length of a `ScopedPageAllocation` is non-zero.

Also add some tests, and add common derives to `ScopedPageAllocation`
and `PageAllocationError` for convenience.

BUG=b:373881381
TEST=cargo xtask check

Change-Id: Iabe1e86d819015afcc76422f3c8f30a266277de7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5937343
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>