xtask: Add flexor build option
Add `--flexor` as an option to xtask build to enable flexor support.
With this option added, the build will fallback to booting flexor if
there is an error while loading the flex kernel. The actual
implementation to boot flexor is in a later CL, this CL just adds the
build option.

BUG=b:350741513
TEST=cargo xtask build --flexor
TEST=cargo xtask check --vm-tests

Change-Id: Id0730ff77b2b2a493e44706a33b4ae22a57c7aba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6001937
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nikhil Gumidelli <nikhilgm@google.com>
Commit-Queue: Nikhil Gumidelli <nikhilgm@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>