crdyboot: avb: Generate a dev path for the root device
Generate the devices prefix path for the PCI device of
the PciRoot of the bootloader.

This only works with PCI devices on bus 0.
It needs to find the proper bus number.

This is the path that the later boot stages will use to
locate the appropriate dynamic partitions for the booted
device.

BUG=b:365802655
TEST=cargo xtask check

Change-Id: I391ea2d7f8d3b649f517f8697a743812f3adbf51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6024687
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>