libcrdy: Add error module
This module provides the `fail_with_fatal_error` macro, which will be
used in crdyshim and crdyboot to end execution when a fatal error
occurs. In following CLs, this will replace the current method of
exiting, which is just a panic.

The actual behavior is more or less the same as the panic hander [1]:
print the error, pause for 10s, then shutdown. The main difference is
that a better error message is printed. It includes the package name and
version, as well as the full error chain (see [2]).

Example of the new output:

```
Boot error in crdyboot-1.0.3: failed to load kernel

Caused by:
    call to LoadKernel failed: 0x100b1000 (VB2_ERROR_LK_INVALID_KERNEL_FOUND)
```

Old output:

```
boot failed: failed to load kernel: call to LoadKernel failed: 0x100b1000 (VB2_ERROR_LK_INVALID_KERNEL_FOUND)
```

[1]: https://github.com/rust-osdev/uefi-rs/blob/68cdc532ed31f2ccab2c9e8f75bb2eef98b4bb7c/uefi/src/helpers/panic_handler.rs#L5
[2]: https://doc.rust-lang.org/std/error/trait.Error.html#method.source

BUG=b:377906805
TEST=cargo xtask check

Change-Id: Ica4eb44179bd77e09b4b355de5a8dace173ff940
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6011943
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>