libcrdy: Make uefi::ScopedProtocol into an opaque struct
Move the contents to a new `ScopedProtocolInner` enum, so that the
internals of `ScopedProtocol` are not visible outside the `uefi` module.

This is a precursor to adding a new unsafe variant; we do not want that
variant to be constructible without calling an `unsafe` method.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: I21e85156e6ddf4500f9adec216df9e0f74a75539
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6021395
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>