crdyshim: Add methods to the Crdyshim trait
This is mostly just shuffling code around.

One notable change is that the PE parsing happens several more times
now. This is done to avoid requiring `Crdyshim` trait methods that
return a value with a lifetime tied to the input. (A parsed PE doesn't
the underlying executable data, but instead references a byte slice.)
Such methods can't be used with automock, so instead we parse the PE
each time it's needed. Parsing the PE is quite quick, so this shouldn't
affect anything visible to the end user.

BUG=b:379917574
TEST=cargo xtask check

Change-Id: Ib314e3d183152d824d61248d42e77a648be954a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6039058
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>