crdyshim: Add FileLoader trait
This allows file loading to be mocked, which will be needed in new unit
tests in a later commit.

This is mostly shuffling code around, but with one notable change to the
implementation: the full path of the file to load is now specified,
e.g. "\efi\boot\crdybootx64.efi", instead of just the file name
portion. Functionally this should behave the same though.

BUG=b:379917574
TEST=cargo xtask check

Change-Id: Id70f8bdaa6d672f52132573e3c8b99035bbbc8b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6039055
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>