crdyboot: Use more realistic BlockIoMedia in tests
Create two BlockIoMedia, one for the hard drive and one for the stateful
partition. These differ only in the media ID and the block length.

In read_blocks, write_blocks, and read_disk, check the media ID to make
sure we're operating on the expected block device.

BUG=b:365161823
TEST=cargo xtask check

Change-Id: Iae83d76686847f98187de15aa67943aad10ce43c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6044113
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>