crdyboot: avb: Use the first PCI node for the sys path
Use the first PCI device node instead of the last.
The init process only checks the first PCI device
in the path so limit the device as such.

Add a test case to ensure the first PCI device in the path
is the one used to generate the path. Only the first
device is used in user space for matching.

Add a case to catch a bug where a USB device in the path
was being used to generate the PCI path and not the PCI
device.

Fix the path generation to include only the first PCI device
in the path.
Fix the generation to match the subtype properly. It was
incorrectly considering a USB node as a PCI node leading
to incorrect paths.

BUG=b:365802655
TEST=boot from USB PCI device, note the first node is used in logs
TEST=boot from a UAS boot stick

Change-Id: Id1bcfc51707db9e9275195fd58a0e240769bebdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6064417
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>