crdyboot: Reduce divergance in vboot/flexor kernel arg handling
Get a regular string from both the vboot and flexor branches, and then
convert to UCS-2 after the vboot/flexor branch.

This is in preparation for making the flexor command-line (slightly)
more dynamic. String manipulation with the `String` type is easier than
with `CString16`.

The kernel command line is now logged for both vboot and flexor.

BUG=b:388328520
TEST=cargo xtask check

Change-Id: I61aa81db0c650174f1719105042c294112f44f78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6154246
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>