Prepare for adding mutable data to the logger implementation
In an upcoming commit, the logger will start keeping a history of recent
log messages. This will require the logger to have access to mutable
state. Since this code is `no_std`, we cannot use `Mutex`. Instead, use
an `AtomicPtr` and carefully document why this is safe.

In this commit there is no actual data stored in the mutable
`LoggerInner`; that will be added in later commits.

BUG=b:392660257
TEST=cargo xtask check --vm-tests
TEST=Observe verbose logs: cargo xtask build && cargo xtask qemu

Change-Id: Ia59bbb77b17e80745f0f15f12983acd1d3244ec6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6216741
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>