Add display_level to Logger
This is in preparation for having the `Logger` always store verbose
logs, but only conditionally print them depending on the
`display_level`.

Prior to this commit we used the max log level of the `log` crate to
filter log messages. Change to always setting the `log` crates max level
to debug (since that is the maximum log level we ever want to see), and
use `display_level` to filter messages in the `Logger`.

BUG=b:392660257
TEST=cargo xtask check --vm-tests
TEST=Observe verbose logs: cargo xtask build && cargo xtask qemu

Change-Id: If6ad4c9d84daa1b5f50917a6d6caf8c21773a018
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6216742
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>