Add history to Logger
This allows the logger to keep track of recent log lines, including
verbose logs that may not be printed to the screen.

For now this history is created but not used. In a later commit it will
be used to print recent logs after a fatal error occurs.

BUG=b:392660257
TEST=cargo xtask check --vm-tests
TEST=Observe verbose logs: cargo xtask build && cargo xtask qemu

Change-Id: If9b5ac3cd7a35338d76e78ec667c54f811a2c550
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6216743
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>