xtask: Drop SignAndUpdateBootloader abstraction
This used to be used in more places, but now it's only used in a couple
places in gen_disk, and doesn't carry its weight in terms of making the
code clearer.

Inline the sign/update operations to the two places that call it.

In addition to making the code simpler, this will help with moving
signature generation out of gen_disk, with the end goal of signatures
being accessible outside the disk image.

BUG=b:402506929
TEST=cargo xtask check --vm-tests

Change-Id: If79e0227d897fa7f06e95a72eb1cf0a067ae221e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6344513
Reviewed-by: Patrick Thompson <ptf@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>