xtask: Sign the bootloaders immediately after building
This makes the signed files easier to access, since they are right next
to the boot files. When testing on a DUT with secure boot enabled (with
custum keys), it's helpful to be able to copy the result of `cargo xtask
build` over to the DUT. And for that to work, the signed files need to
be used.

Note that the signed files are now generated twice; this will be cleaned
up in the next commit.

BUG=b:402506929
TEST=cargo xtask check --vm-tests

Change-Id: Ice1280f9aa5df941b5781dd6e52189c9243ce323
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6343694
Reviewed-by: Patrick Thompson <ptf@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>