xtask: Add deploy action
This copies bootloaders (and signature files) to a DUT's ESP via
SSH. Useful for quickly iterating on changes.

By default, only crdyboot is deployed. Pass `--crdyshim` to also deploy
crdyshim. The reason for this default is that if secure boot is enabled
with custom keys, the signing key for the crdyshim might not match the
enrolled keys, as there is no standard root dev key in vboot_reference.

BUG=b:402506929
TEST=cargo xtask build && cargo xtask deploy <dut>

Change-Id: I56710d8448f8bfbe1d75883c25f3c0c46c2373b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6348616
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Patrick Thompson <ptf@google.com>