Add Aarch64 to the list of architectures in libcrdy
In crdyshim, this requires an update in `get_executable_path`. Per the
spec [1], the aarch64 file name convention is `bootaa64.efi`.

In crdyboot, this requires an update to `execute_linux_kernel`. No
compatibility entry point is needed, so the behavior is the same as
x64_64.

[1]: https://uefi.org/specs/UEFI/2.11/03_Boot_Manager.html#removable-media-boot-behavior

BUG=b:403257806
TEST=cargo xtask check

Change-Id: I12374e4252ce62c4712482aee0a0d1389de1cf7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6352300
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>