libcrdy: Add control over the number of log history lines written
This decouples the number of logs printed after a fatal error from the
number of logs stored by `LogHistory`.

For now this has no effect on behavior; the log history is capped at 20
lines, and the number of lines to print is also set to 20. In a later
commit, the number of lines to store in `LogHistory` will be increased
(to allow storing a higher number of logs in a UEFI var), while the
number of lines to print will be kept unchanged.

BUG=b:405968720
TEST=cargo xtask check

Change-Id: Ie932a8f9535542ced10b785438746838a8ef5264
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6387587
Reviewed-by: Andre Braga <andrebraga@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>