crdyboot: Improve firmware error printing
The errors in FirmwareError are never fatal, so the only place they are
printed is an info log in `update_firmware()`. Since this is using
simple `Display` formatting, error chains are not printed. To ensure
that source errors are printed, remove the `#[source]` annotation and
just format the source error as part of the `Display` string.

BUG=b:397698913
TEST=cargo xtask check

Change-Id: I8249a9422ec86b78df5fc7cf439e24d6b2b112a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6388837
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>