libcrdy: Impl gpt_disk_io::BlockIo for ScopedBlockIo
This allows using gpt_disk_io to read the GPT from a disk via UEFI's
disk block IO protocol. Writes are not needed currently, so an error is
unconditionally returned in write_blocks and flush.

This impl will only be used from the crdyboot package, but it needs to
be implemented in libcrdy due to the orphan rule (ScopedBlockIo is
defined in libcrdy).

BUG=b:397698913
TEST=cargo xtask check

Change-Id: I63366f5e5f2cffbd79f24da53a7d820639577ca0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6431788
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>