xtask: Merge the stateful partition test data into the vboot test disk
This is in preparation for changes to the implementation of the code
that loads firmware update capsules from the stateful partition. Instead
of using a UEFI protocol to access that partition directly, the code
will instead access the entire disk, parse the GPT, and access the
partition at an offset within the disk.

That means that the test code will also have to change: instead of
directly accessing the stateful partition data, it will need to read it
from a full disk image that includes the stateful partition data.

This CL just modifies the code that generates the test data. The actual
pinned test data will be regenerated in a later commit.

BUG=b:397698913
TEST=cargo xtask check

Change-Id: I863c46970f709c585010bdf84ffa93feda41d0f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6431785
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: TingHao Lin <tinghaolin@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>