Update vboot test data
This follows the process described in docs/test_data.md.

In addition to the usual updates to the expected kernel command line,
some updates in the disk tests are needed. Add constants for the start
and length of the stateful partition within the vboot test disk, and use
that to read the partition data from within the larger disk. This is
needed since there is no longer a separate file containing the stateful
partition.

BUG=b:397698913
TEST=cargo xtask check

Cq-Depend: chromium:6431306
Change-Id: I6dff9f3a0e3145f56b2a928ba209bfea3205d1fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6431786
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: TingHao Lin <tinghaolin@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>