xtask: Speculative fix for CoP flakes in TPM test
The CoP tests are sometimes failing on the "successful boot with an
active v1 tpm" test. It's possible that qemu is failing because the
swtpm hasn't fully initialized yet in the slower test environment of
CoP. Try adding a short delay, and see if that resolves the flakes.

BUG=b:407787282
TEST=cargo xtask check

Change-Id: Ic353320950181289e1b0dbd73c0f54e92dd343f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6427319
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>