crdyboot: Switch to gpt_disk_types::GptPartitionEntry
The GptPartitionEntry type is defined in both the uefi crate and the
gpt_disk_types crate. We currently get the `GptPartitionEntry` from the
PartitionInfo protocol, so (prior to this commit) we are only using the
type from the uefi crate. In a later commit, we will switch to using the
gpt_disk_io crate to read the GPT. The gpt_disk_io crate uses the
GptPartitionEntry definition from the gpt_disk_types crate.

To prepare for this transition, switch most of the code to using the
gpt_disk_types definition.

Test code is switched over as well, except for
`DeviceKind::partition_info`, since that's emulating the UEFI API.

BUG=b:397698913
TEST=cargo xtask check

Change-Id: I6ab4fc34bdf8232a1e37f2be9cda51b35dc747a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6442176
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>