crdyboot: Minor cleanups of Hd3 test data
Rename BootDrive::Hd3Mbr to just BootDrive::Hd3. This is consistent with
the other drives, and also prepares for a later commit when this drive
will just be some non-GPT disk, as opposed to specifically MBR.

Also add Hd3 to the DeviceKind enum, and add both Hd3 and its MBR
partition to the list of block IO handles.

BUG=b:397698913
TEST=cargo xtask check

Change-Id: Ifcc6546a184d0e72d5b5c11e453ca3612d53b0dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6442323
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>