crdyboot: Minor cleanups in disk tests
Change the Hd1State partition data to more closely match the STATE
partition in VBOOT_TEST_DISK. In particular, the GUID and partition
length now match the values in gen_disk.rs.

Change the media number in HD1_MEDIA and HD1_STATE_MEDIA to 100 and 101,
respectively. As we add more disk media in later commits, this makes it
easier to see at a glance that 100 corresponds to Hd1, 200 corresponds
to Hd2, etc.

In read_blocks, read from a new `src` variable instead of directly from
VBOOT_TEST_DISK. This is in preparation for making the source disk
conditional in later commits.

BUG=b:397698913
TEST=cargo xtask check

Change-Id: I201827644d4f5a6b42ce5c32e3f799113aa92c46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6442324
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>