crdyboot: Add backing data for Hd2 and Hd3 test disks
Nothing reads this testing data yet, that will happen in later commits.

Use android_test_disk.bin as the backing data for Hd2.

Use an all-zeros array as the backing data for Hd3. This doesn't quite
match the current use of Hd3 as an MBR disk, but in later commits we
will care more about testing a non-GPT disk as opposed to specifically
MBR.

BUG=b:397698913
TEST=cargo xtask check

Change-Id: I7e8bad95680e2a22b5300ea47bc72ac9239543df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/6442326
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>